requirements.txt
torch
torchvision
torchaudio
numpy
pillow
streamlit
openai
python-dotenv
